+++ src/BICO/randomness.cpp
@@ -1,5 +1,0 @@
-#include "randomness.h"
-
-using namespace CluE;
-
-std::mt19937 Randomness::mt19937Generator(time(0));